projects/kailash-irrigation/requirements.txt
# 冈仁波齐天山灌溉系统 - Python 依赖

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP 客户端
requests==2.31.0
httpx==0.25.1

# 数据处理
pydantic==2.5.0
python-dotenv==1.0.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# 时序数据库
influxdb-client==1.38.0

# 异步支持
asyncio-mqtt==0.16.1
aioredis==2.0.1

# 任务调度
apscheduler==3.10.4
celery==5.3.4

# 监控与日志
prometheus-client==0.19.0
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1
